pyproject.toml
[project]
name = "seedeval"
version = "0.1.0"
description = "Evaluation harness for Seedance video-generation outputs."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "imageio-ffmpeg>=0.5.1",
    "numpy>=1.26.0",
    "open-clip-torch>=2.24.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-ulid>=2.7.0",
    "tenacity>=8.3.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[tool.uv]
constraint-dependencies = [
    "torch==2.3.1+cpu",
    "torchvision==0.18.1+cpu",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["seedeval"]
